-반응형, CSS 수정, 리팩토링, 배포

+++ src/components/units/Admin/Admin.queries.ts
@@ -1,7 +1,7 @@
 import { gql } from "@apollo/client";
 
 export const FETCH_CAR_REGISTRATIONS = gql`
-  query fetchCarRegistrations($page: Float!) {
+  query fetchCarRegistrations($page: Int!) {
     fetchCarRegistrations(page: $page) {
       id
       carNumber
